add recurrent to docs
