+++ docs/api_reference/layers/recurrent.rst
@@ -1,0 +1,21 @@
+
+================
+Recurrent Layers
+================
+
+
+.. automodule:: npdl.layers.recurrent
+
+
+.. autoclass:: SimpleRNN
+   :members:
+
+
+.. autoclass:: GRU
+   :memebers:
+
+
+.. autoclass:: LSTM
+   :members:
+
+
